Refactor WeasyPrintClient to value receiver instead of pointer
Updated the WeasyPrintClient struct to use value receiver in various places instead of pointer receiver. This ensures cleaner and less error-prone code in instances where the receiver does not need to mutate the original struct. Also made updates to the Dockerfile to clean up after build to reduce image size.